requirements.txt
ultralytics>=8.0
pyyaml>=6.0
requests>=2.28
